+++ stac_fastapi/tests/api/test_api.py
@@ -54,6 +54,7 @@ ROUTES = {
     "POST /catalogs",
     "GET /catalogs/{catalog_id}",
     "DELETE /catalogs/{catalog_id}",
+    "GET /catalogs/{catalog_id}/children",
     "GET /catalogs/{catalog_id}/collections",
     "POST /catalogs/{catalog_id}/collections",
     "GET /catalogs/{catalog_id}/collections/{collection_id}",